Relocate tests `TestMarshalRecordBodyDateFormat`, `TestFieldDataWithTypedValues`, `TestInvalidTokenSentinelThroughDo`, and `TestURLBuildersEscapeSegments` to improve organization; no changes to test logic.
