+++ errors_test.go
@@ -1,10 +1,7 @@
 package filemaker
 
 import (
-	"context"
 	"errors"
-	"net/http"
-	"net/http/httptest"
 	"testing"
 )
 
@@ -36,21 +33,5 @@ func TestAPIErrorIsMatchesAnyMessage(t *testing.T) {
 	err := error(&APIError{Messages: []Message{{Code: 500, Text: "x"}, {Code: 952, Text: "y"}}})
 	if !errors.Is(err, ErrInvalidToken) {
 		t.Errorf("errors.Is should match a non-primary message carrying the code")
-	}
-}
-
-// With reauth disabled, a 952 from the host should surface as an error that
-// callers can branch on with errors.Is rather than inspecting numeric codes.
-func TestInvalidTokenSentinelThroughDo(t *testing.T) {
-	srv := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		writeJSON(w, `{"response":{},"messages":[{"code":"952","message":"Invalid FileMaker Data API token"}]}`)
-	}))
-	defer srv.Close()
-
-	c := testClient(srv) // reauthOnInvalidToken defaults to false
-	var rb responseBody
-	err := c.do(context.Background(), http.MethodGet, c.baseURL()+"/x", nil, &rb)
-	if !errors.Is(err, ErrInvalidToken) {
-		t.Fatalf("got %v, want errors.Is ErrInvalidToken", err)
 	}
 }
