+++ values_test.go
@@ -93,48 +93,3 @@ func TestDurationMarshalJSON(t *testing.T) {
 		}
 	}
 }
-
-// TestMarshalRecordBodyDateFormat checks that the body carries a dateformats
-// parameter and reformats Date/Timestamp values only when a format is set, and
-// is unchanged (no parameter, US values) when unset.
-func TestMarshalRecordBodyDateFormat(t *testing.T) {
-	fields := FieldData{
-		"DOB":     Date(time.Date(1990, 6, 23, 0, 0, 0, 0, time.UTC)),
-		"Created": Timestamp(time.Date(2026, 6, 23, 14, 5, 0, 0, time.UTC)),
-	}
-	us, iso := DateFormatUS, DateFormatISO
-	cases := []struct {
-		name   string
-		format *DateFormat
-		want   string
-	}{
-		{"unset", nil, `{"fieldData":{"Created":"06/23/2026 14:05:00","DOB":"06/23/1990"}}`},
-		{"US", &us, `{"dateformats":0,"fieldData":{"Created":"06/23/2026 14:05:00","DOB":"06/23/1990"}}`},
-		{"ISO", &iso, `{"dateformats":2,"fieldData":{"Created":"2026-06-23 14:05:00","DOB":"1990-06-23"}}`},
-	}
-	for _, c := range cases {
-		body, err := marshalRecordBody(fields, recordConfig{}, c.format)
-		if err != nil {
-			t.Fatalf("%s: marshalRecordBody: %v", c.name, err)
-		}
-		if string(body) != c.want {
-			t.Errorf("%s:\n got: %s\nwant: %s", c.name, body, c.want)
-		}
-	}
-}
-
-func TestFieldDataWithTypedValues(t *testing.T) {
-	body, err := marshalRecordBody(FieldData{
-		"Active":  Bool(true),
-		"DOB":     Date(time.Date(1990, 6, 23, 0, 0, 0, 0, time.UTC)),
-		"Created": Timestamp(time.Date(2026, 6, 23, 14, 5, 0, 0, time.UTC)),
-		"Name":    "Mark",
-	}, recordConfig{}, nil)
-	if err != nil {
-		t.Fatalf("marshalRecordBody: %v", err)
-	}
-	want := `{"fieldData":{"Active":1,"Created":"06/23/2026 14:05:00","DOB":"06/23/1990","Name":"Mark"}}`
-	if string(body) != want {
-		t.Errorf("got:  %s\nwant: %s", body, want)
-	}
-}
